OpenLIBM: remaining local modifications
The following section is an excerpt of DRAGONFLY.README for this vendor
branch.  Many of these patched issued exist in the current libm as well,
but they are masked with WARNS?=1.  Other fixes adjust for differences
between DragonFly and OpenBSD.  The WARNS level has been raised to 3, and
it could be raised to 4 by fixing "set-but-unused" errors.

The following files have been patched
=====================================
arch/amd64/*.S             (elf note added to all assembly files)
arch/amd64/fenv.c          (#include diff from OpenBSD)
arch/amd64/fenv.h          (#include <sys/cdefs.h> for CDECL)
include/global/math.h      (#include diff, double/float typedefs)
src/e_j0.c                 (full coverage)
src/e_j0f.c                (full coverage, fix sign comparison mismatch)
src/e_j1.c                 (full coverage)
src/e_j1f.c                (full coverage, fix sign comparison mismatch)
src/e_jnf.c                (sign comparison mismatch)
src/e_pow.c                (sign comparison mismatch)
src/e_powf.c               (sign comparison mismatch)
src/e_rem_pio2f.c          (sign comparison mismatch)
src/ld80/e_fmodl.c         (sign comparison mismatch)
src/ld80/s_ceill.c         (sign comparison mismatch)
src/ld80/s_floorl.c        (sign comparison mismatch)
src/ld80/s_nextafterl.c    (strong alias, explicit eval order, always true)
src/ld80/s_remquol.c       (sign comparison mismatch)
src/s_lround.c             (sign comparison mismatch)
src/s_lroundf.c            (sign comparison mismatch)
src/s_nan.c                (#include diff from OpenBSD)
src/w_drem.c               (ansify prototype)

+++ contrib/openbsd_libm/src/e_jnf.c
@@ -181,7 +181,7 @@ ynf(int n, float x)
 	b = y1f(x);
 	/* quit if b is -inf */
 	GET_FLOAT_WORD(ib,b);
-	for(i=1;i<n&&ib!=0xff800000;i++){ 
+	for(i=1;i<n&&(u_int32_t)ib!=0xff800000;i++){
 	    temp = b;
 	    b = ((float)(i+i)/x)*b - a;
 	    GET_FLOAT_WORD(ib,b);
