+++ contrib/openbsd_libm/src/e_rem_pio2f.c
@@ -125,7 +125,7 @@ __ieee754_rem_pio2f(float x, float *y)
 	    fn = (float)n;
 	    r  = t-fn*pio2_1;
 	    w  = fn*pio2_1t;	/* 1st round good to 40 bit */
-	    if(n<32&&(ix&0xffffff00)!=npio2_hw[n-1]) {	
+	    if(n<32&&(ix&0xffffff00)!=(u_int32_t)npio2_hw[n-1]) {
 		y[0] = r-w;	/* quick check no cancellation */
 	    } else {
 	        u_int32_t high;
